app_py/app/requirements.txt
pillow
pandas
streamlit
xlsxwriter
